extract out binary and command module
