+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod ast;
-pub mod cmd;
 pub mod codegen;
 pub mod lexer;
 pub mod parser;
